+++ numpoly/poly_function/__init__.py
@@ -3,6 +3,7 @@ from .call import call
 from .decompose import decompose
 from .derivative import diff, gradient, hessian
 from .isconstant import isconstant
+from .lead_coefficient import lead_coefficient
 from .lead_exponent import lead_exponent
 from .divide import poly_divide, poly_divmod, poly_remainder
 from .set_dimensions import set_dimensions
@@ -10,5 +11,5 @@ from .sortable_proxy import sortable_proxy
 from .tonumpy import tonumpy
 
 __all__ = ("call", "decompose", "diff", "gradient", "hessian", "isconstant",
-           "lead_exponent", "poly_divide", "poly_divmod", "poly_remainder",
+           "lead_coefficient", "lead_exponent", "poly_divide", "poly_divmod", "poly_remainder",
            "set_dimensions", "sortable_proxy", "tonumpy")