+++ numpoly/poly_function/lead_coefficient.py
@@ -1,11 +1,11 @@
-"""Find the lead exponents for each polynomial."""
+"""Find the lead coefficients for each polynomial."""
 import numpy
 import numpoly
 
 
-def lead_exponent(poly, graded=False, reverse=False):
+def lead_coefficient(poly, graded=False, reverse=False):
     """
-    Find the lead exponents for each polynomial.
+    Find the lead coefficients for each polynomial.
 
     As polynomials are not inherently sortable, values are sorted using the
     highest `lexicographical` ordering. Between the values that have the same
@@ -13,7 +13,7 @@ def lead_exponent(poly, graded=False, reverse=False):
 
     Args:
         poly (numpoly.ndpoly):
-            Polynomial to locate exponents on.
+            Polynomial to locate coefficients on.
         graded (bool):
             Graded sorting, meaning the indices are always sorted by the index
             sum. E.g. ``q0**2*q1**2*q2**2`` has an exponent sum of 6, and will
@@ -25,26 +25,25 @@ def lead_exponent(poly, graded=False, reverse=False):
 
     Returns:
         (numpy.ndarray):
-            Integer array with the largest exponents in the polynomials. The
-            shape is ``poly.shape + (len(poly.names),)``. The extra dimension
-            is used to indicate the exponent for the different indeterminants.
+            Array of same shape and type as `poly`, containing all the lead
+            coefficients.
 
     Examples:
         >>> q0, q1 = numpoly.variable(2)
-        >>> numpoly.lead_exponent([1, q0+1, q0**2+q0+1]).T
-        array([[0, 1, 2]])
-        >>> numpoly.lead_exponent(
-        ...     [1, q0, q1, q0*q1, q0**3-1]).T
-        array([[0, 1, 0, 1, 3],
-               [0, 0, 1, 1, 0]])
+        >>> numpoly.lead_coefficient(q0+2*q0**2+3*q0**3)
+        3
+        >>> numpoly.lead_coefficient([1-4*q1+q0, 2*q0**2-q1, 4])
+        array([-4, -1,  4])
 
     """
     poly = numpoly.aspolynomial(poly)
-    shape = poly.shape
-    poly = poly.ravel()
-    out = numpy.zeros(poly.shape+(len(poly.names),), dtype=int)
-    if not poly.size:
+    out = numpy.zeros(poly.shape, dtype=poly.dtype)
+    if not out.size:
         return out
     for idx in numpoly.glexsort(poly.exponents.T, graded=graded, reverse=reverse):
-        out[poly.coefficients[idx] != 0] = poly.exponents[idx]
-    return out.reshape(shape+(len(poly.names),))
+        values = poly.coefficients[idx]
+        indices = values != 0
+        out[indices] = values[indices]
+    if not poly.shape:
+        out = out.item()
+    return out