+++ test/test_poly_function.py
@@ -143,9 +143,20 @@ def test_sortable_proxy():
 
 
 def test_lead_exponent():
+    empty = numpoly.lead_exponent([])
+    assert empty.size == 0
+    assert empty.dtype == int
+    assert empty.shape == (0, 1)
     exponents = numpoly.lead_exponent(POLY1)
     xs, ys = exponents[:, :, 0], exponents[:, :, 1]
     assert numpy.all(xs == [[0, 1, 1, 2], [0, 0, 0, 0],
                             [1, 2, 0, 1], [0, 0, 0, 0]])
     assert numpy.all(ys == [[0, 0, 0, 0], [1, 1, 2, 0],
                             [0, 0, 0, 0], [2, 0, 1, 1]])
+
+
+def test_lead_coefficient():
+   empty = numpoly.lead_coefficient([])
+   assert empty.size == 0
+   assert empty.dtype == int
+   assert empty.shape == (0,)